requirements.txt
albucore==0.0.12
albumentations==1.4.10
annotated-types==0.7.0
astroid==3.2.2
autopep8==2.3.1
certifi==2024.6.2
charset-normalizer==3.3.2
coloredlogs==15.0.1
contourpy==1.2.1
cycler==0.12.1
cython==3.0.10
dill==0.3.8
easydict==1.13
flatbuffers==24.3.25
fonttools==4.53.0
grpcio==1.64.1
grpcio-tools==1.64.1
humanfriendly==10.0
idna==3.7
imageio==2.34.2
isort==5.13.2
joblib==1.4.2
kiwisolver==1.4.5
lazy-loader==0.4
loguru==0.7.2
matplotlib==3.9.0
mccabe==0.7.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
onnx==1.16.1
onnxruntime==1.18.0
onnxruntime-gpu==1.18.0
pillow==10.3.0
prettytable==3.10.0
protobuf==5.27.1
pycodestyle==2.12.0
pydantic==2.7.4
pydantic-core==2.18.4
pylint==3.2.3
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pyyaml==6.0.1
requests==2.32.3
scikit-image==0.24.0
scikit-learn==1.5.0
scipy==1.14.0
sympy==1.12.1
threadpoolctl==3.5.0
tifffile==2024.6.18
tomli==2.0.1
tomlkit==0.12.5
tqdm==4.66.4
urllib3==2.2.2
opencv-python==4.10.0.84